// src/app/models/Pessoa/Pessoa.ts
export class Pessoa {
    id: number = 0;
    dtCriacao: Date;
    nome: string;
    tipoPessoa: number;
    rg: string;
    cpf: string;
    dataNascimento: Date;
    responsavel: string;
    parentesco: string;
    estCivil: number;
    tipoSexo: number;
    empresaId: number;
    excluido: boolean;
}